Change setReflectionProbe to setEnvMap
